university/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Deprecated==1.2.13
Django==4.1.2
django-ajax-selects==2.2.0
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.31
psycopg2==2.9.5
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.5
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
XlsxWriter==3.0.3
